Refactor project to pass cleanjs lint
Not sure how readable chained ternary operators are. The same goes for
using boolean logic short circuiting. Might change lint to allow if/else
expressions for readability. As for no expressions that are not
assigned I'm not sure whether that makes side effects causing code less
or more visible, for example:

// The code bellow is not being assigned, its likely causing a
// side effect
socket.emit('action', action)

// The code bellow is causing a side effect, but its not obvious
// as its being assigned
const emit = socket.emit('action', action)

At the end of the day a thin sliver of the app is going to be doing IO.
Maybe there is not point tricking the linter in this case? Maybe
the answer is to have a folder containing files where all IO is
performed and not running the linter on it?